Backend/requirements.txt
blinker==1.9.0
captcha==0.7.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cors==1.0.1
cryptography==46.0.1
Deprecated==1.2.18
dnspython==2.8.0
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Limiter==3.13
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6

limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
pycparser==2.23
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.15.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
setuptools==80.9.0
tldextract==5.3.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
zope.event==6.0
zope.interface==8.0.1
Faker==19.13.0